# frontend/generated/backend-openapi.json
{
  "components": {
    "schemas": {
...
      "DownloadProgressCompleteResponse": {
        "properties": {
          "status": {
            "const": "complete",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "DownloadProgressCompleteResponse",
        "type": "object"
      },
      "DownloadProgressErrorResponse": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "status": {
            "const": "error",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "error"
        ],
        "title": "DownloadProgressErrorResponse",
        "type": "object"
      },
      "DownloadProgressRunningResponse": {
        "properties": {
          "all_files": {
            "items": {
              "enum": [
                "checkpoint",
                "upsampler",
                "distilled_lora",
                "ic_lora",
                "depth_processor",
                "person_detector",
                "pose_processor",
                "text_encoder",
                "zit"
              ],
              "type": "string"
            },
            "title": "All Files",
            "type": "array",
            "uniqueItems": true
          },
          "completed_files": {
            "items": {
              "enum": [
                "checkpoint",
                "upsampler",
                "distilled_lora",
                "ic_lora",
                "depth_processor",
                "person_detector",
                "pose_processor",
                "text_encoder",
                "zit"
              ],
              "type": "string"
            },
            "title": "Completed Files",
            "type": "array",
            "uniqueItems": true
          },
          "current_downloading_file": {
            "anyOf": [
              {
                "enum": [
                  "checkpoint",
                  "upsampler",
                  "distilled_lora",
                  "ic_lora",
                  "depth_processor",
                  "person_detector",
                  "pose_processor",
                  "text_encoder",
                  "zit"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Downloading File"
          },
          "current_file_progress": {
            "title": "Current File Progress",
            "type": "number"
          },
          "error": {
            "title": "Error",
            "type": "null"
          },
          "expected_total_bytes": {
            "title": "Expected Total Bytes",
            "type": "integer"
          },
          "speed_bytes_per_sec": {
            "title": "Speed Bytes Per Sec",
            "type": "number"
          },
          "status": {
            "const": "downloading",
            "title": "Status",
            "type": "string"
          },
          "total_downloaded_bytes": {
            "title": "Total Downloaded Bytes",
            "type": "integer"
          },
          "total_progress": {
            "title": "Total Progress",
            "type": "number"
          }
        },
        "required": [
          "status",
          "current_downloading_file",
          "current_file_progress",
          "total_progress",
          "total_downloaded_bytes",
          "expected_total_bytes",
          "completed_files",
          "all_files",
          "speed_bytes_per_sec"
        ],
        "title": "DownloadProgressRunningResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelDownloadRequest": {
        "properties": {
          "modelTypes": {
            "items": {
              "enum": [
                "checkpoint",
                "upsampler",
                "distilled_lora",
                "ic_lora",
                "depth_processor",
                "person_detector",
                "pose_processor",
                "text_encoder",
                "zit"
              ],
              "type": "string"
            },
            "title": "Modeltypes",
            "type": "array",
            "uniqueItems": true
          }
        },
        "title": "ModelDownloadRequest",
        "type": "object"
      },
      "ModelDownloadStartResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "status": {
            "const": "started",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message",
          "sessionId"
        ],
        "title": "ModelDownloadStartResponse",
        "type": "object"
      },
...
      "RequiredModelsResponse": {
        "properties": {
          "modelTypes": {
            "items": {
              "enum": [
                "checkpoint",
                "upsampler",
                "distilled_lora",
                "ic_lora",
                "depth_processor",
                "person_detector",
                "pose_processor",
                "text_encoder",
                "zit"
              ],
              "type": "string"
            },
            "title": "Modeltypes",
            "type": "array"
          }
        },
        "required": [
          "modelTypes"
        ],
        "title": "RequiredModelsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LTX-2 Video Generation Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/models/download": {
      "post": {
        "operationId": "route_model_download_api_models_download_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownloadStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route Model Download",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/download/progress": {
      "get": {
        "operationId": "route_download_progress_api_models_download_progress_get",
        "parameters": [
          {
            "in": "query",
            "name": "sessionId",
            "required": true,
            "schema": {
              "title": "Sessionid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/DownloadProgressRunningResponse"
                    },
                    {
                      "$ref": "#/components/schemas/DownloadProgressCompleteResponse"
                    },
                    {
                      "$ref": "#/components/schemas/DownloadProgressErrorResponse"
                    }
                  ],
                  "title": "Response Route Download Progress Api Models Download Progress Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route Download Progress",
        "tags": [
          "models"
        ]
      }
    },
...
    "/api/models/required-models": {
      "get": {
        "operationId": "route_required_models_api_models_required_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "skipTextEncoder",
            "required": false,
            "schema": {
              "default": false,
              "title": "Skiptextencoder",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequiredModelsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route Required Models",
        "tags": [
          "models"
        ]
      }
    },
...
  }
}